requirements.txt
ansible==2.10.7
ansible-base==2.10.5
apache-libcloud==3.2.0
apturl==0.5.2
argcomplete==1.8.1
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.8.2
BTrees==4.9.2
capstone==4.0.2
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colored-traceback==0.3.0
command-not-found==0.3
commonmark==0.9.1
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.2
distro==1.5.0
distro-info==1.0
dnspython==2.1.0
duplicity==0.8.17
ecdsa==0.16.1
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.1
future==0.18.2
httplib2==0.18.1
idna==2.10
impacket==0.9.23
intervaltree==3.1.0
itsdangerous==2.0.1
jeepney==0.6.0
Jinja2==3.0.1
jmespath==0.10.0
kazam==1.4.5
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
ldap3==2.9.1
ldapdomaindump==0.9.3
lockfile==0.12.2
louis==3.16.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.5
netaddr==0.7.19
netifaces==0.10.9
ntlm-auth==1.4.0
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pandas==1.3.2
paramiko==2.7.2
pefile==2021.5.24
persistent==4.7.0
pexpect==4.8.0
Pillow==8.1.2
pipenv==2021.5.29
platformdirs==2.2.0
plumbum==1.7.0
prompt-toolkit==3.0.20
protobuf==3.12.4
psutil==5.8.0
pwncat==0.4.3
pwntools==4.6.0
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycryptodomex==3.10.1
pycups==2.0.1
pycurl==7.43.0.6
pyelftools==0.27
pyfiglet==0.8.post1
Pygments==2.7.1
PyGObject==3.38.0
PyJWT==1.7.1
pykerberos==1.1.14
pymacaroons==0.13.0
pymssql==2.2.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.1
python-debian==0.1.39
python-magic==0.4.20
python-rapidjson==1.4
pytz==2021.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.3.1
qrcode==7.3
reportlab==3.5.66
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
rich==10.9.0
ROPGadget==6.6
rpyc==5.0.1
scp==0.14.0
SecretStorage==3.3.1
selinux==3.1
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.4.0
ssh-import-id==5.11
systemd-python==234
transaction==3.0.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicorn==1.0.2rc3
urllib3==1.26.2
usb-creator==0.3.7
virtualenv==20.7.2
virtualenv-clone==0.5.6
wadllib==1.3.5
wcwidth==0.2.5
Werkzeug==2.0.1
wfuzz==3.1.0
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zc.lockfile==2.0
ZConfig==3.6.0
zdaemon==4.3
ZEO==5.2.3
ZODB==5.6.0
zodbpickle==2.0.0
zodburi==2.5.0
zope.interface==5.4.0
